Fix disposal and add debug
Disposal was actually working in Unity despite the unused pointer argument. Not sure why. Oh well, it is fixed. I did have a crash bug and wrote debug_main to try and debug it. But I actually figured out the issue (it was on the C# side) while writing it and haven't really tried to execute that code. But I figured since I got the cmake for it I might as well commit it.