Rework transforms for QVVS
This is a big breaking change to support QVVS transforms. Currently, it is up to the C# side to coordinate the synchronization of uniform scales with the QVV parts. The uniform scales can be compressed via scalars compression.

I also got rid of AoS vs SoA differentiators. AoS is the clear winner for ACL and QVVS transforms in Unity, partly because of how memory allocation works in Unity, and partly because the number of pose operations that would be accelerated by SoA are less than the number of operations to transpose the data.